requirements.txt
python-chess
stockfish
torch
numpy
